fixed integration test to work with the new changes
